chore(tests): git service test helper
